backend/requirements-dev.txt
# Minimal requirements for local development testing
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7

# Redis and job queue
redis==5.0.1
rq==1.15.1

# Stripe payment processing
stripe==13.0.0

# Video rendering dependencies
playwright==1.40.0
ffmpeg-python==0.2.0

# Google Cloud Storage
google-cloud-storage==2.10.0

# Email delivery
sendgrid==6.10.0

# Rate limiting and security
Flask-Limiter==3.5.0
python-magic==0.4.27

# API documentation
flask-restx==1.3.0

# System monitoring
psutil==5.9.6

# Admin interface
Flask-Admin==1.6.1